+++ R/lifecycle.R
@@ -47,9 +47,6 @@
 #'
 #' * [is_pairlist()], [as_pairlist()], [is_node()], [is_node_list()]
 #'
-#' * [is_definition()], [new_definition()], [is_formulaish()],
-#'   [dots_definitions()]
-#'
 #' * [local_options()], [with_options()], [push_options()],
 #'   [peek_options()], [peek_option()]
 #'