+++ tests/testthat/test-operators.R
@@ -62,12 +62,6 @@ test_that("%@% has replacement version", {
   expect_identical(x, structure(list(), baz = "quux"))
 })
 
-test_that("new_definition() returns new `:=` call", {
-  def <- "foo" ~ "bar"
-  node_poke_car(def, quote(`:=`))
-  expect_identical(new_definition("foo", "bar"), def)
-})
-
 test_that("%@% works with S4 objects (#207)", {
   .Person <- setClass("Person", slots = c(name = "character", species = "character"))
   fievel <- .Person(name = "Fievel", species = "mouse")